CLI Fixes and Reorganization
Fixed multiple issues with the CLI, including fatal issues related to calling the wrong GUI. Also updated commands per last sprint's criticisms.  Also reorganized obsolete classes into their own package, because deleting over 4000K lines of code would be demoralizing.
