+++ src/main/java/velcro/Obsolete/app.java
@@ -1,4 +1,4 @@
-package main.java.velcro.main;
+package main.java.velcro.Obsolete;
 
 import java.awt.EventQueue;
